Structure model classes by packages to improve readability
