+++ src/main/java/ru/hse/se/ucheck/UCheck.java
@@ -1,6 +1,16 @@
 package ru.hse.se.ucheck;
 
-import ru.hse.se.ucheck.models.*;
+import ru.hse.se.ucheck.models.base.Check;
+import ru.hse.se.ucheck.models.base.Coordinates;
+import ru.hse.se.ucheck.models.base.ItemInStore;
+import ru.hse.se.ucheck.models.base.Store;
+import ru.hse.se.ucheck.models.base.item.Tag;
+import ru.hse.se.ucheck.models.cart.CartInStore;
+import ru.hse.se.ucheck.models.cart.ItemInCart;
+import ru.hse.se.ucheck.models.filter.Filter;
+import ru.hse.se.ucheck.models.rating.Rating;
+import ru.hse.se.ucheck.models.rating.Review;
+import ru.hse.se.ucheck.models.sort.SortRule;
 
 import java.time.ZonedDateTime;
 import java.util.List;
