+++ src/main/java/ru/hse/se/ucheck/models/base/Store.java
@@ -1,4 +1,4 @@
-package ru.hse.se.ucheck.models;
+package ru.hse.se.ucheck.models.base;
 
 import java.util.Objects;
