+++ src/main/java/ru/hse/se/ucheck/models/base/item/Measure.java
@@ -1,0 +1,28 @@
+package ru.hse.se.ucheck.models.base.item;
+
+public enum Measure {
+    PIECE,
+    GRAM,
+    KILOGRAM,
+    TON,
+    CENTIMETER,
+    DECIMETER,
+    METER,
+    SQUARE_CENTIMETER,
+    SQUARE_DECIMETER,
+    SQUARE_METER,
+    MILLILITER,
+    LITER,
+    CUBIC_METER,
+    KILOWATT_HOUR,
+    GIGACALORIE,
+    DATE,
+    HOUR,
+    MINUTE,
+    SECOND,
+    KILOBYTE,
+    MEGABYTE,
+    GIGABYTE,
+    TERABYTE,
+    OTHER
+}
