+++ src/main/java/ru/hse/se/ucheck/models/sort/Sortable.java
@@ -1,4 +1,6 @@
-package ru.hse.se.ucheck.models;
+package ru.hse.se.ucheck.models.sort;
+
+import ru.hse.se.ucheck.models.base.Store;
 
 public interface Sortable {
