+++ src/test/java/ru/hse/se/ucheck/test/TestGetFilteredCartInStores.java
@@ -2,7 +2,13 @@ package ru.hse.se.ucheck.test;
 
 import ru.hse.se.ucheck.UCheckException;
 import ru.hse.se.ucheck.UCheckRamImpl;
-import ru.hse.se.ucheck.models.*;
+import ru.hse.se.ucheck.models.base.Check;
+import ru.hse.se.ucheck.models.base.Coordinates;
+import ru.hse.se.ucheck.models.cart.CartInStore;
+import ru.hse.se.ucheck.models.filter.Filter;
+import ru.hse.se.ucheck.models.rating.Review;
+import ru.hse.se.ucheck.models.sort.SortParameter;
+import ru.hse.se.ucheck.models.sort.SortRule;
 
 import java.time.ZoneId;
 import java.time.ZonedDateTime;
